ippserver: partially interact with Mopria Print Service
Get-Printer-Attributes now returns dummy values for all IPP/1.1 required
attributes.  With this in place and dummy responses from Validate-Job
and Print-Job, this server can be added to an Android device via Mopria
Print Service, which will correctly detect the server and send print
data to it.

Note that it's important that the server return "processing" at least
once after hte print job is created.  Otherwise the print job will never
be recognized as having completed.